+++ example2.c
@@ -79,8 +79,8 @@ int main()
         	fgets(line, 128, fp);
         	sscanf(line, "%d/%d/%d %d:%d:%f %lf %lf", &Y, &M, &D, &h, &m, &s, &rtasc, &decl);
         	obs[i][0] = Mjday(Y, M, D, h, m, (double) s);
-        	obs[i][1] = Rad*((double) rtasc);
-        	obs[i][2] = Rad*((double) decl);
+            obs[i][1] = Rad*rtasc;
+            obs[i][2] = Rad*decl;
     	}
 
    	fclose(fp);
