+++ example7.c
@@ -93,8 +93,8 @@ int main()
             exit(3);
         }
         obs[i][0] = Mjday(Y, M, D, h, m, (double) s);
-        obs[i][1] = Rad*((double) rtasc);
-        obs[i][2] = Rad*((double) decl);
+        obs[i][1] = Rad*rtasc;
+        obs[i][2] = Rad*decl;
     }
 
     fclose(fp);
